+++ react_on_rails_pro/CHANGELOG.md
@@ -4,11 +4,22 @@ All notable changes to this project will be documented in this file. Items under
 ## [Unreleased]
 - nothing
 
+## [1.1.0]
+### Added
+- Added `tracing` configuration flag to time server rendering calls
+
+### Changed
+- Default usage of PORT and LOG_LEVEL for the vm-renderer bin file changed to use values RENDERER_PORT and RENDERER_LOG_LEVEL
+
+Above changes in [PR 52](https://github.com/shakacode/react_on_rails_pro/pull/52) by [justin808](https://github.com/justin808).
+
 ## [1.0.0]
 ### Added
 - support for node renderer & fallback renderer
 - support for javascript evaluation caching
 - advanced error handling
 
-[Unreleased]: https://github.com/shakacode/react_on_rails_pro/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/shakacode/react_on_rails_pro/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/shakacode/react_on_rails_pro/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/shakacode/react_on_rails_pro/releases/tag/1.0.0
+